Use Python int in minimum_position

+++ dask_ndmeasure/__init__.py
@@ -636,7 +636,7 @@ def minimum_position(input, labels=None, index=None):
     if not min_1dpos_lbl_rem.ndim:
         min_1dpos_lbl_rem = min_1dpos_lbl_rem[None]
     for i in _pycompat.irange(input.ndim):
-        d = numpy.int64(numpy.prod(input.shape[i + 1:]))
+        d = int(numpy.prod(input.shape[i + 1:]))
         min_pos_lbl.append(min_1dpos_lbl_rem // d)
         min_1dpos_lbl_rem %= d
     min_pos_lbl = dask.array.stack(min_pos_lbl, axis=1)
